size

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,8 @@
 #include<stdio.h>
 /**
  * main - A program that prints the size of various types
- * Return: 0 (succss)
+ *
+ * Return: Always 0 (succss)
  */
 int main(void)
 {
@@ -10,10 +11,10 @@ int b;
 long int c;
 long long int d;
 float f;
-printf("Size of a char: %lu 1 byte(s)\n",(unsigned long)sizeof(a));
-printf("Size of a int: %lu 1 byte(s)\n",(unsigned long)sizeof(b));
-printf("Size of a long int: %lu 1 byte(s)\n",(unsigned long)sizeof(c));
-printf("Size of a long long int: %lu 1 byte(s)\n",(unsigned long)sizeof(d));
-printf("Size of a float: %lu 1 byte(s)\n",(unsigned long)sizeof(f));
+printf("Size of a char: %d 1 byte(s)\n",sizeof(a));
+printf("Size of a int: %d 1 byte(s)\n",sizeof(b));
+printf("Size of a long int: %d 1 byte(s)\n",sizeof(c));
+printf("Size of a long long int: %d 1 byte(s)\n",sizeof(d));
+printf("Size of a float: %d 1 byte(s)\n",sizeof(f));
 return (0);
 }
